Updated check_update() method to only update version option when version is different and to return the filtered variable
Not returning the filtered variable ($update_plugin_option) was causing
an issue with my EDD-powered plugin which hooks to the
"pre_set_site_transient_update_plugins" filter. This filter is typically
only fired when there is a change to the value of the update_plugins
transient. The check_update function was not returning any value so the
pre_set_site_transient filter was firing on every page load.